Custom exceptions moved to api module.
