+++ api/src/main/java/com/tinqinacademy/authenticationservice/api/base/OperationProcessor.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.authenticationservice.api.base;
 
-import com.tinqinacademy.authenticationservice.api.error.Errors;
+import com.tinqinacademy.authenticationservice.api.exceptions.Errors;
 import io.vavr.control.Either;
 
 public interface OperationProcessor<I extends  OperationInput, O extends OperationOutput > {
