+++ api/src/main/java/com/tinqinacademy/authenticationservice/api/exceptions/Errors.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.authenticationservice.api.error;
+package com.tinqinacademy.authenticationservice.api.exceptions;
 
 import lombok.*;
 import org.springframework.http.HttpStatus;
