+++ api/src/main/java/com/tinqinacademy/authenticationservice/api/exceptions/custom/ViolationsException.java
@@ -1,6 +1,6 @@
-package com.tinqinacademy.authenticationservice.core.exceptions.custom;
+package com.tinqinacademy.authenticationservice.api.exceptions.custom;
 
-import com.tinqinacademy.authenticationservice.api.error.Error;
+import com.tinqinacademy.authenticationservice.api.exceptions.Error;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
