+++ core/src/main/java/com/tinqinacademy/authenticationservice/core/exceptions/ExceptionService.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.authenticationservice.core.exceptions;
 
-import com.tinqinacademy.authenticationservice.api.error.Errors;
+import com.tinqinacademy.authenticationservice.api.exceptions.Errors;
 
 public interface ExceptionService {
     Errors handle(Throwable throwable);
